reset DB and change fields to decimal

+++ Marinel/Data/Entities/BookSale.cs
@@ -9,7 +9,7 @@ namespace SchoolDraftWebsite.Data.Entities
         public int Id { get; set; }
         public DateTime Date { get; set; }
         public int NoOfBooksSold { get; set; }
-        public long Revenue { get; set; }
+        public decimal Revenue { get; set; }
         public int InventoryItemId { get; set; }
 
         public virtual InventoryItem InventoryItem { get; set; }
